Rename exclusion_rate to inclusion_rate
This parameter is used in exclusion_cost calculation but in the divisor.
So, calling it inclusionrate better represents its function.

Basically, setting it to a small value (e.g., 0.01, 0.001) would
increase the calculated exclusion costs -- which in turn lets the
solver easily use the vehicles by letting them "exit the depot"
with less number of points.
Setting it to a very higher value (e.g., 0.9, 0.99, 1.1) would make
only very loaded vehicles exit the depot -- so it would be harder
for the solver to use the vehicles.

The parameter has to be a positive integer -- i.e., cannot be zero.

However, it can be greater than 1.0 to account for the fact that
exclusion cost is approximated and we might need more than 1.0
to achive a certain  load level.